+++ internal/repos/github_handler.go
@@ -2,9 +2,12 @@ package repos
 
 import (
 	"context"
+	"fmt"
 	"reflect"
 
 	gh "github.com/google/go-github/v43/github"
+
+	"github.com/sourcegraph/log"
 
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/webhooks"
 	"github.com/sourcegraph/sourcegraph/internal/api"
@@ -15,11 +18,11 @@ import (
 )
 
 type GitHubWebhookAPI struct {
-	Client *github.V3Client
+	client *github.V3Client
 }
 
 func NewGitHubWebhookAPI(client *github.V3Client) *GitHubWebhookAPI {
-	return &GitHubWebhookAPI{Client: client}
+	return &GitHubWebhookAPI{client: client}
 }
 
 func (g *GitHubWebhookAPI) Register(router *webhooks.GitHubWebhook) {
@@ -41,7 +44,28 @@ func (g *GitHubWebhookAPI) handleGitHubWebhook(ctx context.Context, extSvc *type
 
 	repoName := *event.Repo.URL
 	name := api.RepoName(repoName[8:])
-	repoupdater.DefaultClient.EnqueueRepoUpdate(ctx, name)
 
+	resp, err := repoupdater.DefaultClient.EnqueueRepoUpdate(ctx, name)
+	if err != nil {
+		return err
+	}
+
+	log.Scoped("GitHub handler", fmt.Sprintf("Successfully updated: %s", resp.Name))
 	return nil
+}
+
+func (g *GitHubWebhookAPI) CreateSyncWebhook(ctx context.Context, repoName, targetURL, secret string) (int, error) {
+	return g.client.CreateSyncWebhook(ctx, repoName, targetURL, secret)
+}
+
+func (g *GitHubWebhookAPI) ListSyncWebhooks(ctx context.Context, repoName string) ([]github.WebhookPayload, error) {
+	return g.client.ListSyncWebhooks(ctx, repoName)
+}
+
+func (g *GitHubWebhookAPI) FindSyncWebhook(ctx context.Context, repoName string) (int, error) {
+	return g.client.FindSyncWebhook(ctx, repoName)
+}
+
+func (g *GitHubWebhookAPI) DeleteSyncWebhook(ctx context.Context, repoName string, hookID int) (bool, error) {
+	return g.client.DeleteSyncWebhook(ctx, repoName, hookID)
 }
